Create fixCasino.js
Run this file to fix the Casino to always win roulette by betting on 1 with $100000000. Wins $360,000,000 each time. Get up to $10bn before being kicked out of the Casino, then buy a TOR router and run the command "buy -a" in your terminal to get all programs from the darkweb.